build: compile cleanly under MSVC /WX on a non-UTF-8 code page
/utf-8 on every MSVC target so \u escapes in narrow literals are UTF-8
whatever the machine's ANSI code page is (C4566 broke the runtime tests on
cp1252). The coroutine bridge converts interpreter slots with static_cast
for integers (MSVC rejects reinterpret_cast int-to-int), and three /W4
warnings from the recent codegen-brass integration are fixed.

+++ tests/codegen_brass/debug_test.cpp
@@ -208,7 +208,7 @@ TEST_CASE("Line entries map back to source lines and filenames") {
     bool foundFileInCtx = false;
     uint32_t mathFileId = 0;
     for (size_t f = 1; f <= obj->debug_context.file_count(); ++f) {
-        if (obj->debug_context.get_file(f) == filename) {
+        if (obj->debug_context.get_file(static_cast<uint32_t>(f)) == filename) {
             foundFileInCtx = true;
             mathFileId = static_cast<uint32_t>(f);
             break;
